refactor(server): improve DX — slim start/stop, move logging to CLI
- Remove signal handling and startup logging from createServer start()
- Add async stop() that gracefully shuts down HTTP + destroys clients
- Remove destroy() (replaced by stop())
- Move discovery GET / from workspace plugin to createServer
- Remove DEFAULT_SYNC_PORT, inline 3913 in createSyncServer
- Rename createSyncServer destroy() to async stop() for consistency
- CLI serve command owns startup logging and signal handling
